requirements/dev.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements/dev.in
#
amqp==5.1.1
    # via kombu
asgiref==3.6.0
    # via
    #   django
    #   django-structlog
asttokens==2.2.1
    # via stack-data
attrs==22.2.0
    # via
    #   pytest
    #   pytest-mypy
backcall==0.2.0
    # via ipython
billiard==3.6.4.0
    # via celery
black==23.1.0
    # via pytest-black
bumpver==2022.1120
    # via -r requirements/dev.in
celery==5.2.7
    # via
    #   -r requirements/site.in
    #   django-celery-beat
    #   django-structlog
certifi==2022.12.7
    # via
    #   requests
    #   sentry-sdk
charset-normalizer==3.0.1
    # via requests
click==8.1.3
    # via
    #   black
    #   bumpver
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
colorama==0.4.6
    # via
    #   -r requirements/dev.in
    #   bumpver
colorclass==2.2.2
    # via pip-upgrader
coverage[toml]==7.2.1
    # via pytest-cov
decorator==5.1.1
    # via ipython
django==4.1.7
    # via
    #   -r requirements/site.in
    #   django-celery-beat
    #   django-debug-toolbar
    #   django-extensions
    #   django-spaceless-templates
    #   django-storages
    #   django-structlog
    #   django-stubs
    #   django-stubs-ext
django-celery-beat==2.4.0
    # via -r requirements/site.in
django-debug-toolbar==3.8.1
    # via -r requirements/dev.in
django-extensions==3.2.1
    # via -r requirements/site.in
django-ipware==4.0.2
    # via django-structlog
django-spaceless-templates==1.0.2
    # via -r requirements/site.in
django-storages==1.13.2
    # via -r requirements/site.in
django-structlog[celery]==4.1.1
    # via -r requirements/site.in
django-stubs==1.15.0
    # via -r requirements/dev.in
django-stubs-ext==0.7.0
    # via django-stubs
django-timezone-field==5.0
    # via django-celery-beat
docopt==0.6.2
    # via pip-upgrader
exceptiongroup==1.1.0
    # via pytest
executing==1.2.0
    # via stack-data
factory-boy==3.2.1
    # via pytest-factoryboy
faker==17.4.0
    # via factory-boy
filelock==3.9.0
    # via pytest-mypy
flake8==6.0.0
    # via pytest-flake8
freezegun==1.2.2
    # via pytest-freezegun
idna==3.4
    # via requests
inflection==0.5.1
    # via pytest-factoryboy
iniconfig==2.0.0
    # via pytest
ipython==8.11.0
    # via -r requirements/site.in
isort==5.12.0
    # via pytest-isort
jedi==0.18.2
    # via ipython
kombu==5.2.4
    # via celery
lexid==2021.1006
    # via bumpver
matplotlib-inline==0.1.6
    # via ipython
mccabe==0.7.0
    # via flake8
mypy==1.0.1
    # via
    #   django-stubs
    #   pytest-mypy
mypy-extensions==1.0.0
    # via
    #   black
    #   mypy
packaging==23.0
    # via
    #   black
    #   pip-upgrader
    #   pytest
parso==0.8.3
    # via jedi
pathlib2==2.3.7.post1
    # via bumpver
pathspec==0.11.0
    # via black
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pip-upgrader==1.4.15
    # via -r requirements/dev.in
platformdirs==3.0.0
    # via black
pluggy==1.0.0
    # via pytest
prompt-toolkit==3.0.38
    # via
    #   click-repl
    #   ipython
psycopg2-binary==2.9.5
    # via -r requirements/site.in
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pycodestyle==2.10.0
    # via flake8
pyflakes==3.0.1
    # via flake8
pygments==2.14.0
    # via ipython
pymemcache==4.0.0
    # via -r requirements/site.in
pytest==7.2.1
    # via
    #   -r requirements/tests.in
    #   pytest-black
    #   pytest-cov
    #   pytest-django
    #   pytest-factoryboy
    #   pytest-flake8
    #   pytest-freezegun
    #   pytest-isort
    #   pytest-mypy
pytest-black==0.3.12
    # via -r requirements/tests.in
pytest-cov==4.0.0
    # via -r requirements/tests.in
pytest-django==4.5.2
    # via -r requirements/tests.in
pytest-factoryboy==2.5.1
    # via -r requirements/tests.in
pytest-flake8==1.1.1
    # via -r requirements/tests.in
pytest-freezegun==0.4.2
    # via -r requirements/tests.in
pytest-isort==3.1.0
    # via -r requirements/tests.in
pytest-mypy==0.10.3
    # via -r requirements/tests.in
python-crontab==2.7.1
    # via django-celery-beat
python-dateutil==2.8.2
    # via
    #   -r requirements/site.in
    #   faker
    #   freezegun
    #   python-crontab
pytz==2022.7.1
    # via
    #   celery
    #   django-timezone-field
requests==2.28.2
    # via pip-upgrader
sentry-sdk==1.16.0
    # via
    #   -r requirements/site.in
    #   structlog-sentry
six==1.16.0
    # via
    #   asttokens
    #   click-repl
    #   pathlib2
    #   python-dateutil
sqlparse==0.4.3
    # via
    #   django
    #   django-debug-toolbar
stack-data==0.6.2
    # via ipython
structlog==22.3.0
    # via
    #   django-structlog
    #   structlog-sentry
structlog-sentry==2.0.0
    # via -r requirements/site.in
terminaltables==3.1.10
    # via pip-upgrader
toml==0.10.2
    # via
    #   bumpver
    #   pytest-black
tomli==2.0.1
    # via
    #   black
    #   coverage
    #   django-stubs
    #   mypy
    #   pytest
traitlets==5.9.0
    # via
    #   ipython
    #   matplotlib-inline
types-pytz==2022.7.1.2
    # via django-stubs
types-pyyaml==6.0.12.8
    # via django-stubs
typing-extensions==4.5.0
    # via
    #   django-stubs
    #   django-stubs-ext
    #   mypy
    #   pytest-factoryboy
tzdata==2022.7
    # via django-celery-beat
urllib3==1.26.14
    # via
    #   requests
    #   sentry-sdk
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.6
    # via prompt-toolkit

# The following packages are considered to be unsafe in a requirements file:
# setuptools
